Plugin isolation: invoke tool-result truncation directly, not as a plugin (#39352 follow-up)
`tool-result-truncate` was a passthrough plugin over a kernel the daemon
needed anyway: `context/tool-result-truncation.ts` imported
`truncateToolResultText` straight out of the plugin directory, the last
host->plugin edge on the Phase 1 scoreboard for this plugin. Same shape as
history-repair, so the same playbook applies.

- Kernel relocated into `context/tool-result-truncation.ts`, which already
  held `truncateToolResultsAcrossHistory` over the same primitive. One
  daemon module now owns per-result truncation and the across-history pass.
- The agent loop tail-drops each oversized tool result inline at its
  `post-tool-use` chokepoint, before the hook chain — the position the
  default hook occupied by registering first, now unconditional. A single
  result blowing the context window is a hard provider rejection, never a
  per-conversation opt-in, so this no longer sits behind the enabled-plugin
  gate: a workspace that disabled `default-tool-result-truncate` gets the
  bound back.
- Spooling still runs ahead of truncation, so a spooled file keeps the
  tool's full output and truncation sees the stub.

Deleted: the plugin directory (hook, manifest, `terminal.ts`), its barrel
registration, and its reverse-guard baseline entry. Phase 1
`tool-result-truncate: 1 -> 0`; it was already Phase 2 clean.

Coverage: `tool-result-truncate-hook` is replaced by
`agent-loop-tool-result-truncation`, which drives the loop through an
oversized tool result and asserts the bound lands in the provider-bound
history and the emitted `tool_result` event, and that a registered
`post-tool-use` hook observes the already-truncated content. Kernel
correctness stays covered by `tool-result-truncation`.

Also corrects the `history-repair` module docstring, which still described
hooks and a `./hooks/` directory removed in #39366.
